Add an option to streamer configuration named 'use_shell' which causes us to use create_subprocess_shell instead of create_subprocess_exec in run_process.
This is useful when we want to run a streamer cmd that involves piping the output of one program to another program - e.g., using ffmpeg or gstreamer in conjunction with another program to get h264 frames to janus via memsink/etc.

There are also changes to kill_process, as otherwise the only program that would be killed when use_shell=True would be the top-level shell, with the children process continuing to run. I tried to preserve the behavior/logic for when there is only a single process to be the exact same, though I'll admit I didn't 100% follow the logic.