Potential fix for CMD key woes in macOS
Forces a "key-up" event to be generated when combining CMD key with some other key.  For example, the following sequence will result in the "R" key being held down on the remote side...

-Press down CMD
-Press down R
-Release R

The "R" will not register as being released until the "CMD" key is released.  This is sub-optimal with most things (especially things like FN-CMD-TAB, etc.).