yamlconf: support validating "dynamic" keys
If the schema definition at a given level has a `Dynamic` marker key,
validate every key that is not otherwise present in the schema against
the definition of that key.

Further down the line, we can make unexpected extra keys illegal,
but for that we need to sanitize a;; existing uses of dynamic content.